% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{AUC.landmark}
\alias{AUC.landmark}
\title{Calculate AUC for Landmark Prediction}
\usage{
AUC.landmark(t0, tau, data, short = TRUE, weight = NULL)
}
\arguments{
\item{t0}{Landmark time.}

\item{tau}{Prediction window.}

\item{data}{Data frame.}

\item{short}{Logical.}

\item{weight}{Optional weights.}
}
\description{
Calculate AUC for Landmark Prediction
}
