% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{coef.landpred_model_continuous}
\alias{coef.landpred_model_continuous}
\title{Extract Coefficients from Landpred Continuous Model}
\usage{
\method{coef}{landpred_model_continuous}(object, t_s = NULL, ...)
}
\arguments{
\item{object}{A landpred_model_continuous object.}

\item{t_s}{Optional short-term event time.}

\item{...}{Additional arguments.}
}
\value{
A named vector of coefficients.
}
\description{
Extracts coefficients. If \code{t_s} is provided, it fits the short-term GLM
and returns its coefficients.
}

