% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{Ghat.FUN}
\alias{Ghat.FUN}
\title{Estimate Survival Function}
\usage{
Ghat.FUN(tt, data, type = "fl", weight.given)
}
\arguments{
\item{tt}{Time points.}

\item{data}{Data frame.}

\item{type}{Type of estimator.}

\item{weight.given}{Optional weights.}
}
\description{
Estimate Survival Function
}
