% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{helper.si}
\alias{helper.si}
\title{Helper Function for AUC}
\usage{
helper.si(yy, FUN, Yi, Vi = NULL)
}
\arguments{
\item{yy}{Vector 1.}

\item{FUN}{Comparison operator.}

\item{Yi}{Vector 2.}

\item{Vi}{Optional weights.}
}
\description{
Helper Function for AUC
}
