% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{mse.BW}
\alias{mse.BW}
\title{Optimize Bandwidth for Kernel Smoothing}
\usage{
mse.BW(data, t0, tau, h, folds = 3, reps = 2)
}
\arguments{
\item{data}{The data frame.}

\item{t0}{The landmark time.}

\item{tau}{The prediction window.}

\item{h}{The bandwidth to test.}

\item{folds}{Number of cross-validation folds.}

\item{reps}{Number of repetitions.}
}
\value{
The MSE.

The MSE.
}
\description{
Calculates the Mean Squared Error (MSE) for a given bandwidth to help select the optimal bandwidth.
}
