% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{mse_cv}
\alias{mse_cv}
\title{Calculate MSE for Bandwidth Selection using Cross-Validation}
\usage{
mse_cv(
  bw,
  landpred_obj,
  t0,
  tau,
  transform = identity,
  reps = 50,
  train_prop = 0.66
)
}
\arguments{
\item{bw}{The bandwidth to test.}

\item{landpred_obj}{The landpred object.}

\item{t0}{The landmark time.}

\item{tau}{The prediction window.}

\item{transform}{Transformation function for short-term covariate.}

\item{reps}{Number of repetitions.}

\item{train_prop}{Proportion of data to use for training.}
}
\value{
The Mean Squared Error.
}
\description{
Calculate MSE for Bandwidth Selection using Cross-Validation
}
\keyword{internal}
