% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{optimize.mse.BW}
\alias{optimize.mse.BW}
\title{Optimize Bandwidth (Legacy)}
\usage{
optimize.mse.BW(
  data,
  t0,
  tau,
  h.grid = seq(0.01, 2, length = 50),
  folds = 3,
  reps = 2
)
}
\arguments{
\item{data}{Data frame.}

\item{t0}{Landmark time.}

\item{tau}{Prediction window.}

\item{h.grid}{Grid of bandwidths.}

\item{folds}{Number of folds.}

\item{reps}{Number of repetitions.}
}
\description{
Optimize Bandwidth (Legacy)
}
