% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{predict.landpred_model_continuous}
\alias{predict.landpred_model_continuous}
\title{Predict Method for Landpred Continuous Model}
\usage{
\method{predict}{landpred_model_continuous}(object, newdata = NULL, type = "response", ...)
}
\arguments{
\item{object}{A landpred_model_continuous object.}

\item{newdata}{New data frame containing covariates and short-term event info.}

\item{type}{Type of prediction (default "response").}

\item{...}{Additional arguments}
}
\value{
A vector of predicted probabilities.
}
\description{
Predicts the probability of the event occurring given new data.
}
