% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landpred.R
\name{predict.landpred_model_discrete}
\alias{predict.landpred_model_discrete}
\title{Predict Method for Discrete Landpred Model}
\usage{
\method{predict}{landpred_model_discrete}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A landpred_model_discrete object.}

\item{newdata}{Optional new data.}

\item{...}{Additional arguments.}
}
\value{
Predicted probabilities.
}
\description{
Predicts probabilities using the discrete landpred model.
}
