% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{print.landpred_model_continuous}
\alias{print.landpred_model_continuous}
\title{Print Method for Landpred Continuous Model}
\usage{
\method{print}{landpred_model_continuous}(x, ...)
}
\arguments{
\item{x}{A landpred_model_continuous object.}

\item{...}{Additional arguments.}
}
\description{
Prints the continuous landpred model results.
}
