% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landpred.R
\name{print.landpred_model_discrete}
\alias{print.landpred_model_discrete}
\title{Print Method for Discrete Landpred Model}
\usage{
\method{print}{landpred_model_discrete}(x, ...)
}
\arguments{
\item{x}{A landpred_model_discrete object.}

\item{...}{Additional arguments.}
}
\description{
Prints the discrete landpred model results.
}
