% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landpred.R
\name{print.landpred_object}
\alias{print.landpred_object}
\title{Print Method for Landpred Object}
\usage{
\method{print}{landpred_object}(x, ...)
}
\arguments{
\item{x}{A landpred_object.}

\item{...}{Additional arguments.}
}
\description{
Prints a summary of the landpred object.
}
