% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{Prob2.k.t}
\alias{Prob2.k.t}
\title{Calculate Probability Component with Kernel Smoothing}
\usage{
Prob2.k.t(t, t0, tau, data.use, bandwidth, covariate.value, weight = NULL)
}
\arguments{
\item{t}{Time points.}

\item{t0}{Landmark time.}

\item{tau}{Prediction window.}

\item{data.use}{Data frame.}

\item{bandwidth}{Bandwidth.}

\item{covariate.value}{Covariate value.}

\item{weight}{Optional weights.}
}
\description{
Calculate Probability Component with Kernel Smoothing
}
