% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{Prob2}
\alias{Prob2}
\title{Calculate Probability Component (No Kernel)}
\usage{
Prob2(t0, tau, data, covariate.value, weight = NULL)
}
\arguments{
\item{t0}{Landmark time.}

\item{tau}{Prediction window.}

\item{data}{Data frame.}

\item{covariate.value}{Covariate value.}

\item{weight}{Optional weights.}
}
\description{
Calculate Probability Component (No Kernel)
}
