% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{prob2.single}
\alias{prob2.single}
\title{Calculate Single Probability Component}
\usage{
prob2.single(K, W2i, Xi.long, tau, Di.short, Xi.short, Zi, t0, covariate.value)
}
\arguments{
\item{K}{Kernel weights.}

\item{W2i}{IPCW weights.}

\item{Xi.long}{Long term event times.}

\item{tau}{Prediction window.}

\item{Di.short}{Short term event indicators.}

\item{Xi.short}{Short term event times.}

\item{Zi}{Covariates.}

\item{t0}{Landmark time.}

\item{covariate.value}{Covariate value.}
}
\description{
Calculate Single Probability Component
}
