% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{summary.landpred_model_continuous}
\alias{summary.landpred_model_continuous}
\title{Summary Method for Landpred Continuous Model}
\usage{
\method{summary}{landpred_model_continuous}(object, t_s = NULL, ...)
}
\arguments{
\item{object}{A landpred_model_continuous object.}

\item{t_s}{Optional short-term covariate time.}

\item{...}{Additional arguments.}
}
\description{
Prints a summary of the model, including coefficients and standard errors.
}
