% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landpred.R
\name{summary.landpred_object}
\alias{summary.landpred_object}
\title{Summary Method for Landpred Object}
\usage{
\method{summary}{landpred_object}(object, ...)
}
\arguments{
\item{object}{A landpred_object.}

\item{...}{Additional arguments.}
}
\description{
Prints a detailed summary of the landpred object.
}
