% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{var.fun}
\alias{var.fun}
\title{Estimate Variance of Coefficients}
\usage{
var.fun(t, data.v, tau, s, h, vmat, Ainv, weight = NULL, transform = identity)
}
\arguments{
\item{t}{The target time for the short-term covariate (usually t_s).}

\item{data.v}{The data frame used for estimation.}

\item{tau}{The landmark time.}

\item{s}{The prediction window.}

\item{h}{The bandwidth.}

\item{vmat}{A matrix of perturbation weights.}

\item{Ainv}{The inverse information matrix from the model fit.}

\item{weight}{Optional weights.}

\item{transform}{Transformation function.}
}
\value{
A list containing the estimated standard errors for the intercept and slopes.
}
\description{
Estimates the variance of the coefficients for the short-term GLM using
a perturbation resampling method.
}
\keyword{internal}
