% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{Wi.FUN}
\alias{Wi.FUN}
\title{Calculate Inverse Probability of Censoring Weights}
\usage{
Wi.FUN(data, t0, tau, weight.given = NULL)
}
\arguments{
\item{data}{Data frame.}

\item{t0}{Landmark time.}

\item{tau}{Prediction window.}

\item{weight.given}{Optional weights.}
}
\description{
Calculate Inverse Probability of Censoring Weights
}
