% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{coefficient_se}
\alias{coefficient_se}
\title{Calculate Standard Errors for Coefficients}
\usage{
coefficient_se(model, t_s = NULL, samples = 200)
}
\arguments{
\item{model}{A landpred_model_continuous object.}

\item{t_s}{The time of the short-term covariate measurement.}

\item{samples}{The number of resampling iterations.}
}
\value{
A named vector of standard errors.

}
\description{
Calculates standard errors for the coefficients of the landpred model.
If \code{t_s} is provided, it uses the perturbation resampling method.
Otherwise, it returns the standard errors from the GLM.
}
\keyword{internal}
