% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{landpred}
\alias{landpred}
\title{Create a Landpred Object}
\usage{
landpred(formula, data, discrete = FALSE)
}
\arguments{
\item{formula}{A formula object with a Surv object on the LHS and covariates on the RHS.}

\item{data}{The data frame.}

\item{discrete}{Logical, whether to use the discrete method (legacy).}
}
\value{
A landpred_object.
}
\description{
Parses the formula and data to create a landpred object used for landmark prediction.
}
